% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{ndiff}
\alias{ndiff}
\title{Calculate NDVI}
\usage{
ndiff(b1, b2)
}
\arguments{
\item{b1}{value or vector with reflectance in the first (eg. NIR for NDVI) wavelenght}

\item{b2}{value or vector with reflectance in the second (eg. red for NDVI) wavelenght}
}
\value{
NDVI value
}
\description{
This function caculates the normalized difference vegetation index.
}
